my_strchr.c
With this code I have created a function that searches for the first occurrence of the character c (an unsigned char) in the string pointed to by the argument str. The terminating null character is considered to be part of the string. The function is returning a pointer pointing to the last matching character, or null if no match was found.